added signInPage object for login, added a verification method for login seperate to login method, cleaned up some formatting
