+++ src/test/java/ServiceObjects/HomePage.java
@@ -1,9 +1,6 @@
 package ServiceObjects;
 
-import io.restassured.RestAssured;
 import io.restassured.http.Cookies;
-import io.restassured.internal.common.assertion.Assertion;
-import io.restassured.response.Response;
 import org.junit.jupiter.api.Assertions;
 
 public class HomePage extends BasePage {
@@ -52,29 +49,4 @@ public class HomePage extends BasePage {
         Assertions.assertEquals("Search results", xmlParseByAttribute("class", "h2"));
     }
 
-    public void login(String userName, String password) {
-        getCookiesFilter();
-        Response response = httpRequest
-                .contentType("multipart/form-data")
-                .queryParam("controller", "authentication")
-                .queryParam("back", "my-account")
-                .accept("text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9")
-                .redirects().follow(true)
-                .multiPart("token", "8b3ee4f1cba54a1031be9bb7879e9dad")
-                .multiPart("back", "my-account")
-                .multiPart("email", userName)
-                .multiPart("password", password)
-                .multiPart("submitLogin", "1")
-                .when()
-                .post();
-        response.getBody().prettyPrint();
-        response = httpRequest.get().andReturn();
-        System.out.println(response.getStatusCode());
-        Response response1 = httpRequest.get().andReturn();
-        response1.getBody().prettyPrint();
-        Assertions.assertEquals("My account", response1.htmlPath().getString("html.head.title"));
-
-
-
-    }
 }
